% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfafunctions96.R
\name{calc_preference}
\alias{calc_preference}
\title{calc_preference: calculate the preference in 2 choice MFA using
dataframe returned from comsumption_from_xml()}
\usage{
calc_preference(data_list, equation = "difference")
}
\arguments{
\item{data_list:}{dataframe generated using consumption_from_xml()}
}
\description{
calc_preference: calculate the preference in 2 choice MFA using
dataframe returned from comsumption_from_xml()
}
\examples{

}
\keyword{}
