% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfafunctions96.R
\name{consumption_from_xml}
\alias{consumption_from_xml}
\title{consumption_from_xml: calculate consumption from pairs of plates}
\usage{
consumption_from_xml(
  imp_xml,
  vol = 10,
  matrix_type = "custom",
  layout_matrix = NULL,
  calc_duration = FALSE,
  soln_index = c("S", "N", "N", "C"),
  sample_ID
)
}
\arguments{
\item{imp_xml:}{An internal xml document imported via xml2::read_xml().

vol: The initial volume dispensed into the wells of the microplate.
Default vol = 10.

matrix_type: A character variable indicating the type of built-in
layout used to organize the 96-well culture plate. Default matrix_type = "custom".

For custom matrix layouts, matrix_type = "custom".
layout_matrix A matrix containing a user-provided layour matrix for
sample in the 96-well culture plate. Default layout_matrix = NULL.

Example layout matrix with C = CSB, E = Evaporation, M = Male, F = Female
des_matrix = matrix(c("C", "E", "M", "M", "M", "M", "M", "F", "F", "F", "F", "F",
"C", "E", "M", "M", "M", "M", "M", "F", "F", "F", "F", "F",
"C", "E", "M", "M", "M", "M", "M", "F", "F", "F", "F", "F",
"C", "E", "M", "M", "M", "M", "M", "F", "F", "F", "F", "F",
"C", "E", "M", "M", "M", "M", "M", "F", "F", "F", "F", "F",
"C", "E", "M", "M", "M", "M", "M", "F", "F", "F", "F", "F",
"C", "E", "M", "M", "M", "M", "M", "F", "F", "F", "F", "F",
"C", "E", "M", "M", "M", "M", "M", "F", "F", "F", "F", "F"),
nrow = 8,
ncol = 12,
byrow = TRUE)

calc_duration: A logical variable indicating whether to calculate the time
interval between spectrophotometer readings. Default calc_duration = FALSE.

soln_index: A 4 element list indicating the location of solutions
in the 4 wells from which the flies drink.
Default soln_index = c("S", "N", "N", "C"), with S = Sucrose, N = Nothing,
and C = Cocaine. The order indicates the following layout:
S N
N C}
}
\description{
consumption_from_xml: calculate consumption from pairs of plates
}
\examples{
imp_xml = xml2::read_xml(xml_doc_path)

layout_matrix = matrix(c("C", "E", "M", "M", "M", "M", "M", "F", "F", "F", "F", "F",
                         "C", "E", "M", "M", "M", "M", "M", "F", "F", "F", "F", "F",
                         "C", "E", "M", "M", "M", "M", "M", "F", "F", "F", "F", "F",
                         "C", "E", "M", "M", "M", "M", "M", "F", "F", "F", "F", "F",
                         "C", "E", "M", "M", "M", "M", "M", "F", "F", "F", "F", "F",
                         "C", "E", "M", "M", "M", "M", "M", "F", "F", "F", "F", "F",
                         "C", "E", "M", "M", "M", "M", "M", "F", "F", "F", "F", "F",
                         "C", "E", "M", "M", "M", "M", "M", "F", "F", "F", "F", "F"),
                         nrow = 8,
                         ncol = 12,
                         byrow = TRUE)


con_data = consumption_from_xml(imp_xml)

Output:
ID               Proj    Line  Sex   Meth  Group Rep   Solution      Cons(uL)
DGRP_0112_2MFA   DGRP    0112   M    2MFA    M    1        S        1.28602090
DGRP_0112_2MFA   DGRP    0112   M    2MFA    M    1        C        0.84100040
DGRP_0112_2MFA   DGRP    0112   M    2MFA    M    2        S        0.60428090
DGRP_0112_2MFA   DGRP    0112   M    2MFA    M    2        C        1.02102090
DGRP_0112_2MFA   DGRP    0112   M    2MFA    F    1        S        1.81433180
DGRP_0112_2MFA   DGRP    0112   M    2MFA    F    1        C        2.17056277
DGRP_0112_2MFA   DGRP    0112   M    2MFA    F    2        S        1.18248175
DGRP_0112_2MFA   DGRP    0112   M    2MFA    F    2        C        1.38156028

}
\keyword{absorbance,}
\keyword{consumption}
\keyword{layout,}
\keyword{read,}
