% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfafunctions96.R
\name{csb_adjust}
\alias{csb_adjust}
\title{csb_adjust:}
\usage{
csb_adjust(data_list, csb_label = "C")
}
\arguments{
\item{data_list:}{dataframe generated using consumption_from_xml()

evap_label: character used to indicate evaporation wells in the layout
matrix. Default evap_label = "E".}
}
\description{
csb_adjust:
}
\examples{

}
\keyword{}
