% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfafunctions96.R
\name{evap_adjust}
\alias{evap_adjust}
\title{evap_adjust:}
\usage{
evap_adjust(data_list, evap_label = "E")
}
\arguments{
\item{data_list:}{dataframe generated using consumption_from_xml()

evap_label: character used to indicate evaporation wells in the layout
matrix. Default evap_labe = "E".}
}
\description{
evap_adjust:
}
\examples{

}
\keyword{}
