% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfafunctions96.R
\name{link_xml_to_layout}
\alias{link_xml_to_layout}
\title{link_xml_to_layout:}
\usage{
link_xml_to_layout(layout_dir, xml_dir)
}
\arguments{
\item{data_list:}{dataframe generated using consumption_from_xml()

evap_label: character used to indicate evaporation wells in the layout
matrix. Default evap_label = "E".}
}
\description{
link_xml_to_layout:
}
\examples{

}
\keyword{}
